Refactor mygrad.nnet.layers
