+++ src/mygrad/tensor_manip/tiling/ops.py
@@ -2,7 +2,7 @@ from typing import Optional, Sequence, Union
 
 import numpy as np
 
-from mygrad.nnet.layers.utils import sliding_window_view
+from mygrad.nnet.layers.operations.utils import sliding_window_view
 from mygrad.operation_base import Operation
 from mygrad.tensor_base import Tensor
