+++ tests/nnet/layers/test_conv.py
@@ -11,7 +11,7 @@ from pytest import raises
 
 import mygrad as mg
 from mygrad import Tensor
-from mygrad.nnet.layers import conv_nd
+from mygrad.nnet.layers.operations import conv_nd
 
 from ...utils.numerical_gradient import numerical_gradient_full
 from ...wrappers.uber import backprop_test_factory, fwdprop_test_factory
