+++ tests/nnet/layers/test_maxpool.py
@@ -2,7 +2,7 @@ import numpy as np
 from numpy.testing import assert_allclose
 from pytest import raises
 
-from mygrad.nnet.layers import max_pool
+from mygrad.nnet.layers.operations import max_pool
 from mygrad.tensor_base import Tensor
